refactor: move console-flows to components/console
Co-locate with its only consumers (step-flow.tsx,
flow-completion-modal.tsx). Removes now-empty config/ directory.
